Connect state mashine with view and database
The state mashine as it is now is fully functional. Based on the current state a player is in the score will be handled differently.
Missing two things: 1. States now are not saved in the database which means upon reload of the site  the player will start again with DoubleIn. 2. The very first currentPlayer is not right in the view.